backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic-settings
pandas
scikit-learn
xgboost
redis
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
python-multipart